# swagger.yaml
swagger: '2.0'
info:
  title: PLC Gateway API
  description: 三菱PLCとMCプロトコルで通信するためのGateway API。Copilot Studioなどの外部ツールやMCPクライアントから利用可能です。
  contact:
    name: PLC Gateway API開発チーム
    url: https://github.com/your-org/plc-gateway
    email: support@your-org.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: 1.2.0
paths:
  /api/read:
    post:
      summary: 単一デバイス読み取り
      description: 指定したPLCデバイスから値を読み取ります
      operationId: read_device
      tags:
      - Device Read
      produces:
      - application/json
      responses:
        '200':
          description: Successful Response
          schema:
            $ref: '#/components/schemas/ReadResponse'
        '422':
          description: Validation Error
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
      consumes:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/components/schemas/ReadRequest'
  /api/read/{device}/{addr}/{length}:
    get:
      summary: 単一デバイス読み取り (GET)
      description: URLパスパラメータを使用してPLCデバイスから値を読み取ります
      operationId: read_device_get
      tags:
      - Device Read
      produces:
      - application/json
      responses:
        '200':
          description: Successful Response
          schema:
            $ref: '#/components/schemas/ReadResponse'
        '422':
          description: Validation Error
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: device
        in: path
        required: true
        description: デバイス種別（D, W, R, ZR, X, Y, M）
        type: string
      - name: addr
        in: path
        required: true
        description: デバイスアドレス（10進数または16進数）
        type: integer
        format: int32
      - name: length
        in: path
        required: true
        description: 読み取り長（ワード数またはビット数）
        type: integer
        format: int32
      - name: plc_host
        in: query
        required: false
        description: PLCのコンピューター名またはIPアドレス（省略時は環境変数使用）
        type: string
      - name: ip
        in: query
        required: false
        description: PLCのIPアドレス（後方互換性のため残存、plc_hostを推奨）
        type: string
      - name: port
        in: query
        required: false
        description: PLCのポート番号（省略時は環境変数使用）
        type: integer
  /{device}/{addr}/{length}:
    get:
      summary: Api Read Get Compat
      description: 後方互換性のため、/api/プレフィックスなしでもアクセス可能
      operationId: read_device_compat
      tags: []
      produces:
      - application/json
      responses:
        '200':
          description: Successful Response
          schema:
            $ref: '#/components/schemas/ReadResponse'
        '422':
          description: Validation Error
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: device
        in: path
        required: true
        description: デバイス種別（D, W, R, ZR, X, Y, M）
        type: string
      - name: addr
        in: path
        required: true
        description: デバイスアドレス（10進数または16進数）
        type: integer
        format: int32
      - name: length
        in: path
        required: true
        description: 読み取り長（ワード数またはビット数）
        type: integer
        format: int32
      - name: plc_host
        in: query
        required: false
        description: PLCのコンピューター名またはIPアドレス（省略時は環境変数使用）
        type: string
      - name: ip
        in: query
        required: false
        description: PLCのIPアドレス（後方互換性のため残存、plc_hostを推奨）
        type: string
      - name: port
        in: query
        required: false
        description: PLCのポート番号（省略時は環境変数使用）
        type: integer
  /api/batch_read:
    post:
      summary: 複数デバイス一括読み取り
      description: 複数のPLCデバイスを一括で読み取ります。MCプロトコルのrandomread機能を活用した効率的な読み取りを行います。
      operationId: batch_read_devices
      tags:
      - Batch Operations
      produces:
      - application/json
      responses:
        '200':
          description: Successful Response
          schema:
            $ref: '#/components/schemas/BatchReadResponse'
        '422':
          description: Validation Error
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
      consumes:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        schema:
          $ref: '#/components/schemas/BatchReadRequest'
  /api/batch_read_status:
    get:
      summary: バッチ読み取り機能の状態確認
      description: バッチ読み取り機能のサポート状況と制限事項を取得します
      operationId: api_batch_read_status_api_batch_read_status_get
      tags:
      - System Status
      produces:
      - application/json
      responses:
        '200':
          description: Successful Response
          schema: {}
  /api/openapi/json:
    get:
      summary: OpenAPI仕様ファイル（JSON）のダウンロード
      description: OpenAPI仕様をJSON形式でダウンロードします
      operationId: download_openapi_json_api_openapi_json_get
      tags:
      - OpenAPI Schema
      produces:
      - application/json
      responses:
        '200':
          description: Successful Response
          schema: {}
  /api/openapi/yaml:
    get:
      summary: OpenAPI仕様ファイル（YAML）のダウンロード
      description: OpenAPI仕様をYAML形式でダウンロードします
      operationId: download_openapi_yaml_api_openapi_yaml_get
      tags:
      - OpenAPI Schema
      produces:
      - application/json
      responses:
        '200':
          description: Successful Response
          schema: {}
  /api/openapi/swagger:
    get:
      summary: Swagger 2.0仕様ファイル（JSON）のダウンロード
      description: Swagger 2.0仕様をJSON形式でダウンロードします（Copilot Studio用）
      operationId: download_swagger_json_api_openapi_swagger_get
      tags:
      - OpenAPI Schema
      produces:
      - application/json
      responses:
        '200':
          description: Successful Response
          schema: {}
  /api/openapi/swagger/yaml:
    get:
      summary: Swagger 2.0仕様ファイル（YAML）のダウンロード
      description: Swagger 2.0仕様をYAML形式でダウンロードします（Copilot Studio用）
      operationId: download_swagger_yaml_api_openapi_swagger_yaml_get
      tags:
      - OpenAPI Schema
      produces:
      - application/json
      responses:
        '200':
          description: Successful Response
          schema: {}
  /api/openapi/status:
    get:
      summary: OpenAPI仕様生成機能の状態確認
      description: OpenAPI仕様ファイル生成機能の状態とサポート情報を取得します
      operationId: api_openapi_status_api_openapi_status_get
      tags:
      - OpenAPI Schema
      produces:
      - application/json
      responses:
        '200':
          description: Successful Response
          schema: {}
definitions:
  BatchReadRequest:
    properties:
      devices:
        items:
          type: string
        type: array
        title: Devices
      plc_host:
        anyOf:
        - type: string
        - type: 'null'
        title: Plc Host
      ip:
        anyOf:
        - type: string
        - type: 'null'
        title: Ip
      port:
        anyOf:
        - type: integer
        - type: 'null'
        title: Port
    type: object
    required:
    - devices
    title: BatchReadRequest
  BatchReadResponse:
    properties:
      results:
        items:
          $ref: '#/components/schemas/DeviceReadResult'
        type: array
        title: Results
      total_devices:
        type: integer
        title: Total Devices
      successful_devices:
        type: integer
        title: Successful Devices
    type: object
    required:
    - results
    - total_devices
    - successful_devices
    title: BatchReadResponse
  DeviceReadResult:
    properties:
      device:
        type: string
        title: Device
      values:
        items: {}
        type: array
        title: Values
      success:
        type: boolean
        title: Success
      error:
        anyOf:
        - type: string
        - type: 'null'
        title: Error
    type: object
    required:
    - device
    - values
    - success
    title: DeviceReadResult
    description: デバイス読み取り結果
  HTTPValidationError:
    properties:
      detail:
        items:
          $ref: '#/components/schemas/ValidationError'
        type: array
        title: Detail
    type: object
    title: HTTPValidationError
  ReadRequest:
    properties:
      device:
        type: string
        title: Device
        default: D
      addr:
        type: integer
        title: Addr
      length:
        type: integer
        title: Length
      plc_host:
        anyOf:
        - type: string
        - type: 'null'
        title: Plc Host
      ip:
        anyOf:
        - type: string
        - type: 'null'
        title: Ip
      port:
        anyOf:
        - type: integer
        - type: 'null'
        title: Port
    type: object
    required:
    - addr
    - length
    title: ReadRequest
  ReadResponse:
    properties:
      values:
        items:
          type: integer
        type: array
        title: Values
    type: object
    required:
    - values
    title: ReadResponse
    description: PLCデバイス読み取りレスポンス
  ValidationError:
    properties:
      loc:
        items:
          anyOf:
          - type: string
          - type: integer
        type: array
        title: Location
      msg:
        type: string
        title: Message
      type:
        type: string
        title: Error Type
    type: object
    required:
    - loc
    - msg
    - type
    title: ValidationError
tags:
- name: Device Read
  description: PLCデバイス読み取り操作
- name: Batch Operations
  description: バッチ読み取り操作
- name: System Status
  description: システム状態確認
- name: OpenAPI Schema
  description: OpenAPI仕様ファイルの取得
schemes:
- http
host: localhost:8000
basePath: ''